feat: 4단계 txUserService 구현 및 추상화
